# docs/sad/03-design/openapi.yaml
openapi: 3.0.3
info:
  title: JPLearn Platform API
  version: 0.1.0
  description: 'Foundation v1. No grammar, flashcards, or L1 translation channels.


    Contract authority is this file (OpenAPI 3.0.3), not FastAPI-generated 3.1 JSON.

    Semantic normalized diff (ADR-003 D9): must match HTTP status, required/nullability,

    security, operationId, x-jplearn-fr, and error schema. Allowlist: 3.0↔3.1 nullable

    vs type union, generated component names, servers, info.


    Docs UI (`/docs`, `/redoc`, `/openapi.json`): off on staging and prod; local dev
    may

    enable. Contract tests read this file in git — do not depend on a public OpenAPI
    URL.

    '
servers:
- url: https://api.staging.jplearn.local
paths:
...
  /me:
    get:
      operationId: getMe
      tags:
      - Identity
      x-jplearn-fr:
      - FR-ID-002
      - FR-ID-004
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /staff/catalog/{id}/media:
    post:
      operationId: uploadMedia
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CMS-001
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Media asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAssetStaff'
  /media/{id}:
    get:
      operationId: streamMedia
      tags:
      - Media
      x-jplearn-fr:
      - FR-CMS-003
      - FR-CMS-004
      description: 'MP4/audio bytes. Auth is Bearer JWT **or** HMAC query `exp`+`sig`
        (TTL 1h) so `<video src>` can play without Authorization. Secret = MEDIA_SIGNING_SECRET
        or JWT_SECRET.

        '
      security:
      - bearerAuth: []
      - signedQuery: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: exp
        schema:
          type: integer
        description: Unix seconds expiry (required if no Bearer)
      - in: query
        name: sig
        schema:
          type: string
          pattern: ^[a-f0-9]{64}$
      responses:
        '200':
          description: Media stream
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '401':
          description: Missing or invalid JWT/signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Http401Error'
        '404':
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
  /media/{id}/hls/{file}:
    get:
      operationId: streamHls
      tags:
      - Media
      x-jplearn-fr:
      - NFR-PERF-002
      description: 'HLS manifest (.m3u8) and segments (.ts/.m4s); MP4 /media/{id}
        stays the fallback. Auth is Bearer JWT **or** HMAC query `exp`+`sig` (same
        as GET /media/{id}) so hls.js can fetch segments without Authorization. Manifest
        rewrite copies query onto relative URIs.

        '
      security:
      - bearerAuth: []
      - signedQuery: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: file
        required: true
        schema:
          type: string
          pattern: ^[A-Za-z0-9._-]+$
      - in: query
        name: exp
        schema:
          type: integer
        description: Unix seconds expiry (required if no Bearer)
      - in: query
        name: sig
        schema:
          type: string
          pattern: ^[a-f0-9]{64}$
      responses:
        '200':
          content:
            application/vnd.apple.mpegurl: {}
            video/mp2t: {}
            video/iso.segment: {}
        '400':
          description: Invalid or unsupported file name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '401':
          description: Missing or invalid JWT/signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Http401Error'
        '404':
          description: Asset or HLS file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
  /staff/media/{id}/hls:
    post:
      operationId: registerHls
      tags:
      - CMS
      x-jplearn-fr:
      - FR-CMS-001
      - NFR-PERF-002
      description: Marks an asset HLS-ready after scripts/transcode-hls.sh wrote the
        bundle to storage/hls/{id}/.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '201':
          description: hls_url registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAssetStaff'
        '400':
          description: HLS manifest missing on disk
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '403':
          description: Not teacher or admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
...
  /me/recommendations:
    get:
      tags:
      - Recommendations
      summary: Get Recommendations
      operationId: getRecommendations
      security:
      - bearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          description: Maximum number of recommendations to return
          default: 10
          title: Limit
        description: Maximum number of recommendations to return
      responses:
        '200':
          description: Recommended items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationsResponsePublic'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Http401Error'
      x-jplearn-fr:
      - UC-L16
      - UC-L17
      - FR-WAT-001
      - FR-NEG
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    signedQuery:
      type: apiKey
      in: query
      name: sig
      description: HMAC-SHA256 signature for media streaming (with exp timestamp)
  schemas:
...
    HttpError:
      type: object
      description: Standard JSON error body (400, 403, 404, 500) per ADR-005.
      required:
      - statusCode
      - message
      - error
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
          - type: string
          - type: array
            items:
              type: string
        error:
          type: string
    Http401Error:
      type: object
      description: Distinct 401 Unauthorized JSON error body per ADR-005.
      required:
      - statusCode
      - message
      properties:
        statusCode:
          type: integer
          enum:
          - 401
        message:
          type: string
    MediaAssetStaff:
      type: object
      required:
      - id
      - catalog_item_id
      - storage_key
      - playback_url
      - mime
      properties:
        id:
          type: string
          format: uuid
        catalog_item_id:
          type: string
        storage_key:
          type: string
        playback_url:
          type: string
        hls_url:
          type: string
          nullable: true
        mime:
          type: string
        measured_duration_ms:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
        source_sha256:
          anyOf:
          - type: string
            pattern: ^[a-f0-9]{64}$
          - type: 'null'
        hls_bundle_sha256:
          anyOf:
          - type: string
            pattern: ^[a-f0-9]{64}$
          - type: 'null'
...
    User:
      type: object
      required:
      - id
      - email
      - roles
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        roles:
          type: array
          items:
            type: string
            enum:
            - learner
            - teacher
            - admin
...
    RecommendationsResponsePublic:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RecommendedItemPublic'
          type: array
          title: Items
        strategy_version:
          type: string
          title: Strategy Version
      additionalProperties: false
      type: object
      required:
      - items
      - strategy_version
      title: RecommendationsResponsePublic
    RecommendedItemPublic:
      properties:
        catalog_item_id:
          type: string
          title: Catalog Item Id
        media_type:
          type: string
          title: Media Type
        topic_id:
          type: string
          title: Topic Id
        duration_seconds:
          type: integer
          title: Duration Seconds
        ci_level:
          type: integer
          title: Ci Level
        reason:
          type: string
          enum:
          - continue_series
          - preferred_topic
          - same_level
          - editor_pick
          title: Reason
        title_jp:
          anyOf:
          - type: string
          - type: 'null'
          title: Title Jp
        series_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Series Id
        series_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Series Title
      additionalProperties: false
      type: object
      required:
      - catalog_item_id
      - media_type
      - topic_id
      - duration_seconds
      - ci_level
      - reason
      title: RecommendedItemPublic
...